requirements.txt
pandas
numpy
scikit-learn
transformers
sentence-transformers
stanza
tensorflow-macos
# tf_keras